Update tests to wait for idle event stream.
